put SuperMegaiPadView into its own file
